aidevteam/requirements.txt
pydantic>=2.0.0
langchain-core>=0.1.0
langgraph>=0.0.10
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
openai>=1.0.0
python-dotenv>=1.0.0
